Style 패널, Line 서브 패널 추가
